feat: 修复高级筛选器、优化 UI 交互并新增递归视图与创建功能 (修复所有编译错误)
1. 筛选系统：重构为 QSortFilterProxyModel 架构，实现“置顶优先”排序规则，修正评级/颜色筛选失效。
2. 导航系统：NavPanel 新增桌面/此电脑入口，采用懒加载 QStandardItemModel 架构；完善 QDir 包含。
3. UI/UX：筛选行增高至 24px，修复颜色环截断；状态栏实时显示过滤后计数。
4. 交互：通过焦点判定修复重命名方向键失效；新增 Layers 递归按钮及新建文件菜单。
5. 编译修复：全面补充 QDir, QStandardItem 等头文件；处理 size_t 与 int 间的显式转换，消除 x64 下的歧义报错。

+++ src/ui/ContentPanel.cpp
@@ -927,7 +927,7 @@ void GridItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& opti
 
     int baseIconSize = option.decorationSize.width();
     if (baseIconSize <= 0) baseIconSize = 64; 
-    int iconDrawSize = baseIconSize * 0.65; 
+    int iconDrawSize = static_cast<int>(baseIconSize * 0.65);
     
     int ratingH = 12;
     int nameH = 16;
